Modular app filter code & add filter state to assessments
